+++ web/src/components/Router/Router.tsx
@@ -8,7 +8,7 @@ import RunDetail from 'pages/RunDetail';
 import Settings from 'pages/Settings';
 import Test from 'pages/Test';
 import Transaction from 'pages/Transaction';
-import TransactionRunTrigger from 'pages/TransactionRunTrigger';
+import TransactionRunOverview from 'pages/TransactionRunOverview';
 import TransactionRunAutomate from 'pages/TransactionRunAutomate';
 import AutomatedTestRun from 'pages/AutomatedTestRun';
 import Env from 'utils/Env';
@@ -30,8 +30,8 @@ const Router = () => (
       <Route path="/test/:testId/run" element={<AutomatedTestRun />} />
 
       <Route path="/transaction/:transactionId" element={<Transaction />} />
-      <Route path="/transaction/:transactionId/run/:runId" element={<TransactionRunTrigger />} />
-      <Route path="/transaction/:transactionId/run/:runId/trigger" element={<TransactionRunTrigger />} />
+      <Route path="/transaction/:transactionId/run/:runId" element={<TransactionRunOverview />} />
+      <Route path="/transaction/:transactionId/run/:runId/overview" element={<TransactionRunOverview />} />
       <Route path="/transaction/:transactionId/run/:runId/automate" element={<TransactionRunAutomate />} />
 
       <Route path="*" element={<Navigate to="/" />} />
