+++ web/src/pages/TransactionRunOverview/TransactionRunOverview.tsx
@@ -2,10 +2,10 @@ import TransactionRunLayout from 'components/TransactionRunLayout';
 import withAnalytics from 'components/WithAnalytics/WithAnalytics';
 import Content from './Content';
 
-const TransactionRunTrigger = () => (
+const TransactionRunOverview = () => (
   <TransactionRunLayout>
     <Content />
   </TransactionRunLayout>
 );
 
-export default withAnalytics(TransactionRunTrigger, 'transaction-details-trigger');
+export default withAnalytics(TransactionRunOverview, 'transaction-details-overview');
