+++ web/src/pages/TransactionRunOverview/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line no-restricted-exports
-export {default} from './TransactionRunTrigger';
+export {default} from './TransactionRunOverview';
